+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+=== Head
+
+* Skip nodes that do not have `source` properties.
+
 === 2.3.2 (February 14, 2015)
 
 * Prevent failure if a PostCSS node lacks a `source` property.